pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "adenoma-agent"
version = "0.1.0"
description = "ROI/patch-level pathology segmentation agent MVP for adenoma slides"
requires-python = ">=3.8"
dependencies = [
  "PyYAML>=6.0",
  "Pillow>=9.0",
  "numpy>=1.20",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
